refactor(backend): extract PlatformTarget to separate module
Move PlatformTarget struct to its own file for better code organization:
- Create src/backend/platform_target.rs with the struct definition
- Add comprehensive tests for PlatformTarget functionality
- Update mod.rs imports and remove unused Platform import
- Maintain existing API through public re-export

This improves modularity and makes the codebase easier to maintain.
